Everything working, including tests
Still testing with local client, also going to write more tests before merging back to master

+++ spec/db/postgres/integration/truncate.ts
@@ -3,7 +3,9 @@ import { models } from './models';
 export default async function truncate() {
     const promises = [];
     for (const model in models) {
-        promises.push(models[model].destroy({ where: {}, force: true }));
+        if (models[model].isInitialized) {
+            promises.push(models[model].destroy({ where: {}, force: true }));
+        }
     }
     return await Promise.all(promises);
 }
